+++ views/webrtc-debug/src/components/JoinScreen/index.ts
@@ -1,0 +1,3 @@
+import JoinScreen from "./JoinScreen";
+
+export default JoinScreen;